Extract NamedObjects and TracksAccounts concern
